change storybook architecture
