+++ ui/tools/storybook/stories/pixano/components/modals/ConfirmModal.stories.ts
@@ -15,15 +15,25 @@
 
 // Imports
 import type { Meta, StoryObj } from "@storybook/svelte";
-import { LoadingModal } from "@pixano/core";
+import { ConfirmModal } from "@pixano/core";
 
 const meta = {
-  title: "Components/Core/Modals/LoadingModal",
-  component: LoadingModal,
+  title: "Pixano/Components/Modals/ConfirmModal",
+  component: ConfirmModal,
   tags: ["autodocs"],
-} satisfies Meta<LoadingModal>;
+} satisfies Meta<ConfirmModal>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-export const BasicLoadingModal: Story = {};
+export const BasicConfirmModal: Story = {
+  args: { message: "Some message" },
+};
+
+export const ConfirmModalWithDetails: Story = {
+  args: { message: "Some message", details: "Some details" },
+};
+
+export const ConfirmModalWithAlternativeAction: Story = {
+  args: { message: "Some message", alternativeAction: "Some alternative action" },
+};
