+++ ui/tools/storybook/stories/pixano/components/ui/Histogram.stories.ts
@@ -18,7 +18,7 @@ import type { Meta, StoryObj } from "@storybook/svelte";
 import { Histogram } from "@pixano/core";
 
 const meta = {
-  title: "Components/Core/UI/Histogram",
+  title: "Pixano/Components/UI/Histogram",
   component: Histogram,
   tags: ["autodocs"],
 } satisfies Meta<Histogram>;
